// rebuildr-backend/src/apis/types/postnord/track-shipment-url.ts
export interface LinksResponse {
  faults?: Fault[];
  url?: string;
}

/**
 * ParamValues
 * Key/Value information about errors
 */
export interface ParamValues {
  param?: string;
  value?: string;
}

/**
 * Fault
 * Model for API Fault
 */
export interface Fault {
  explanationText?: string;
  paramValues?: ParamValues[];
  faultCode?: string;
}